pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemsieve"
version = "0.1.0"
description = "Gmail inbox intelligence pipeline — extract gems from your inbox"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "typer[all]>=0.9",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "spacy>=3.7",
    "anthropic>=0.40",
    "httpx>=0.27",
    "openpyxl>=3.1",
    "python-dateutil>=2.8",
    "dacite>=1.8",
]

[project.optional-dependencies]
crew = [
    "crewai>=1.9",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "starlette-admin>=0.14",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
gemsieve = "gemsieve.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/gemsieve"]
